refactor: move deck list container to page
